[refact]: make print_scene directory
